refactor(_apiref): split _ast into _docstring_sections and _preview, retire qast alias
